Create wp-contact-form-7-recaptcha
WIth this code you can hide the recaptcha iframe till the email field is not empty and valid. Once the recaptcha is OK the disabled attribute on the submit button is removed and form can be proccesed.
The scripts requires jquery registered/enqueed, and also will run only if WPCF7 installed.